Move ecc point serialization to PublicKey
